Tpetra: Matrix Market test: Fixed construction of Map.
My test was constructing the Map by making a raw array containing
the list of global indices owned by each process.  I then made an
ArrayView of the list and passed it into Map's nonmember constructor.

Map was perfectly fine, but the test code was not.  I was making the
raw array in an inner scope and assigning it to the ArrayView.  Thus,
the ArrayView wasn't valid.  This explains why it worked in a debug
build but not a release build: in the release build, the compiler
was taking legitimate optimizations to make that raw scope-local
array invalid outside its scope.

My previous commit shows my effort to see whether Map was at fault.
As far as I can tell, Map was not at fault.  However, my previous
commit improves type safety and improves performance of the Map
constructor anyway.