Ifpack2: Relaxation: Fixed initialization of L1Eta_ and Condest_.
Eric Phipps pointed out today that L1Eta_ and Condest_ are both
magnitude_type, but I was assigning a scalar_type value to them.
This would break the cases where scalar_type is complex, or where
scalar_type is a Stokhos automatic differentiation type, for
example.  This commit fixes the problem.