Ported the algorithm that reuses Cijk entries to product Legendre.
I ported the version of the multiply algorithm that loads multiple FEM
columns into shared memory and reuses Cijk values across those columns
to the product Legendre data structure.  I also modified the kernel to
not load the offsets into shared memory, as this was reducing the number
of columns I could load.  It is still a little slower than the original
CRS version, and I think that is because of treating diagonals and
off-diagonals separately.