Tpetra: CrsGraph, CrsMatrix: Refactored local replace/sumInto.
I refactored CrsGraph::transformValues into two separate methods:
transformLocalValues and transformGlobalValues.  This made the methods
easier to read, so that it's easier for us to do performance tuning.
I added a corresponding CrsMatrix::transformLocalValues method, and
refactored CrsMatrix's sumIntoLocalValues and replaceLocalValues
methods to call transformLocalValues.  This removed a lot of
duplicated code and should also help with performance tuning.