Tpetra: CrsGraph: Added findLocalIndex that takes hint & column view.
Thanks to Eric Phipps for contributing a three-argument version of
findLocalIndex that takes a hint for the column index offset.  I
added a hint argument to the version of findLocalIndex that takes a
view of the column indices for that row.  I also made
transformLocalValues call this four-argument version of
findLocalIndex.

The thing to try next would be to optimize the binary search in
findLocalIndex (all versions, if the graph is sorted).  It currently
uses std::equal_range.  This is a legit binary search, but it does
two comparisons (a < b && !(b < a)) per entry examined.  It would be
better to do a single == comparison per entry examined.  Also, Matt
Bettencourt suggested two optimizations:

1. Just use a linear search if the row has fewer than 50 entries.

2. Unroll the search loop by 8.  (This makes sense if the local
   ordinal type is int.  You probably want the unroll length times
   the size of the local ordinal type to fit in one cache line.)

These got him a 30-40% performance improvement in his branch of
Epetra.  (He suggested them to the Epetra64 developers, but they
didn't get integrated, and so they aren't in Epetra yet.)

Eric Phipps tells me that Epetra and Tpetra are within 2% of each
other at this point for sumIntoLocalValues, so we might be able to
defer these optimizations in Tpetra.