Kokkos: DefaultArithmetic: Assign(A,B) now checks for pointer equality.
Assign(A,B) does a deep copy of B into A.  It now checks whether the
raw pointers of A and B are equal.  If they are, it doesn't do the
copy.

Assign(A,B,whichVectors) now does the analogous thing for each
column at a time.

Build/Test Cases Summary
Enabled Packages: KokkosClassic, Tpetra
0) MPI_DEBUG => passed: passed=82,notpassed=0 (22.63 min)
1) SERIAL_RELEASE => passed: passed=73,notpassed=0 (3.98 min)
Other local commits for this build/test group: 859e652