Teuchos: Comm: Refactored reduceAll & added ireceive specializations.
The implementation of the reduceAll specializations (that bypass
serialization and go straight to MPI, if possible) had a lot of
duplicated code.  I factored out the duplicated code by adding a
(currently invisible to users) traits class MpiTypeTraits to get the
MPI_Datatype for an instance of a type T, a generic implementation
that uses the traits class, and a "fall-back" (that uses
serialization) if the Comm is neither an MpiComm nor a SerialComm.

Analogous to the reduceAll specializations, I also added ireceive
specializations for the same template parameters.  (I have not yet
added send and isend specializations; once I do this, the most
important specializations for Tpetra performance will be complete.)
This required exposing the stored "tag" in MpiComm -- a hack that
the original author of the class included, either because they
didn't understand MPI_Comm_dup semantics, or because MPI_Comm_dup
was expensive on their target platform.  (The stored tag is used
to implement copying an MpiComm instance, so that the new instance
has a different communication space than the original.  I have heard
that some platforms have limits on the number of communicators that
can exist at once.)