Kokkos,Ifpack2: Chebyshev: Added kernel for thresholding 1/d[i].
Some Ifpack2 methods, such as Chebyshev and the relaxations, require
a vector of the inverses of the local diagonal entries of the sparse
matrix.  Tpetra::Vector already has a reciprocal() method for that.
The problem is that these Ifpack2 methods let users provide a minimum
diagonal value, which replaces diagonal entries that are too small.
Ifpack2 used to solve this problem by accessing the data directly and
writing a sequential loop.  This is a setup bottleneck when running
Tpetra with many threads, esp. in a NUMA environment.  It's also ugly
code.

I've fixed this by implementing a Kokkos::DefaultArithmetic kernel for
the "thresholded in-place reciprocal" operation.  Ifpack2 can use this
kernel on the local data (the Kokkos::MultiVector returned by the
Tpetra::Vector's getLocalMVNonConst() method).  Ifpack2::Chebyshev now
uses this kernel.  Eventually we can port Relaxation and the other
methods in Ifpack2 to use this kernel.