Ifpack2: Chebyshev: Improved public documentation and code formatting.
The documentation now includes a detailed description of all the
accepted parameters, along with a concordance with the accepted
parameters of Ifpack and ML.

The apply() method now respects the "mode" argument.  Of course the
implementation currently wouldn't work correctly with a nonsymmetric
matrix, but we should still respect the argument.