Ifpack2: Chebyshev: Removed dead code that calls AztecOO with Tpetra.
Originally, it was hoped that Tpetra would be compatible with AztecOO.
This hope never materialized.  Chebyshev had some old code that would
call AztecOO's CG with Tpetra objects.  It was protected by an #ifdef,
but that functionality never worked as far as I know, so I removed the
old code completely and replaced it with an unconditional throw of
std::logic_error.

CG is easy enough to implement that if somebody really wants this
functionality, we should just put a CG implementation in place right
there.  Otherwise, I'll leave it alone for now.