MueLu: now parameter list names can differ from the generated data
In the old system, it was assumed that when you call Input(level, name)
in any factory, the name would coincide with the required generated
data. Therefore, in GetValidParameterList() you set the parameter entry
to be this name.
However, this approach does not work well when a two
level factory requests the same data but from two different levels. In
this case, one might want/need to have two different parameters for two
levels. For instance, the ConstraintFactory in Energy-Minimization
requires nullspaces from both fine and coarse levels. However, we cannot
use the same factory for them, as for fine level we would need a
rebalanced nullspace, but for coarse we would need not rebalanced
(otherwise, we enter the loop).
Thefore, I added a new Input(level, name, param_name) which allows to
have param_name different from name.