Ifpack2: Imitation of Ifpack::Chebyshev replicates Ifpack2 exactly.
This occurs with the test problem if I set lambdaMax to the exact
max eigenvalue, and let lambdaMin and eigRatio take their default
values (or set eigRatio to 30, which is Ifpack's and Ifpack2's
default).

This tells me that Ifpack and Ifpack2 really do implement the same
algorithm; they just might differ a bit in the settings of various
parameters.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => passed: passed=21,notpassed=0 (8.76 min)
2) SERIAL_DEBUG => passed: passed=21,notpassed=0 (14.81 min)