Ifpack2: Fixed setParameters in new Chebyshev implementation.
For many parameters, the new Chebyshev implementation accepts the
Ifpack and ML name for the parameter, if one exists, as well as the
Ifpack2 name.

This is good, but the setParameters() method used to populate the
input ParameterList with all of those different names for the same
parameter, even if they weren't supplied.  This could confuse users.
Now the method only populates the default parameters that it should
populate.