Kokkos: Improved public documentation of LocalSparseOps example.
EmptySparseKernel is an example of a local sparse matrix operations
implementation.  It's a trivial example (the methods do nothing), but
the interface is right.  I plan to push more operations down from the
Tpetra level to the Kokkos level -- for example, computing the explicit
transpose of a sparse matrix.  Improving the documentation of this
class will help me with that.