KokkosArray:  Experimenting with hand-written SSE2/AVX instructions.
In one of the KokkosArray UQ mat-vec kernels, I have manually put in
SSE2/AVX instructions to see what performance improvement there might
be.  I did this by rewriting the kernel to use a blocked inner loop
where the inner blocks are encapsulated in a TinyVector class.  This
class is then specialized for double and length 2 and 4 vectors
incorporating the SSE2/AVX instructions.  While this works and generates
good looking vector instructions, its performance improvement is negligable.

Also modified the CrsProductTensor to use View<> directly instead of
CrsArray, to allow experimenting with Left and Right Layouts and how
they impact the vector instructions.