Update backwards compatibility testing to work on the 11.0 chain
The backwards compatibility was working off of an old copy of the 10.x chain
which is no longer needed for verification. Moving it to the 11.0 branch as this
is the first release chain that we will be rigorously checking backwards
compatibility for specific packages.

Also added an option to the script that builds an install of Trilinos. I'm not
sure how this was over looked when I initially wrote that script, but it was
only building with a single process which is a waste of resources for the
nightly build which generally blocks of 4 or more cores for building.