Setting install testing to experimental
Due to an error the installation testing hasn't been run for a while. The error
should be fixed, however, there are a number of installation issues that need to
be addressed. I am setting this to experimental temporarily to prevent spamming
the developers while I address the backlog of installation issues.