Galeri: Fix installation issues
There were a number of headers supporting Xpetra that weren't being added to the
library and as such weren't being installed. I added one header to the library
that may be a test only header, however, if I were to just flag the src-xpetra
directory as something that needs to be an include directory for testing then we
would miss any files that are added to that directory that actually need to be
installed. There were a number that were missing in this round so I felt the
cost of possibly installing something we don't need to out weighed the risk of
potentially missing files we do need to install.