TrilinosCouplings: Fixed build errors in my previous commit.
Oops, TrilinosCouplings is secondary stable and therefore doesn't
build in the checkin tests, even if you've changed a file in it.
To test it, you have to set up an "extra" configuration to test,
and test it using the --extra-builds=... argument to the check-in
test script.

Build/Test Cases Summary
Enabled Packages: Epetra, Tpetra, TrilinosCouplings
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=97,notpassed=0 (2.65 min)