Fixing invalid SET() statement.
My project has just disabled the TPINode type with:

  SET(KokkosClassic_ENABLE_ThreadPool OFF CACHE BOOL "")

This else statement in Kokkos that I am fixing here was supposedly designed
the set the default but it was written wrong.  This never would have
configured for anyone and therefore I am pretty sure I am the first person to
have actually tested this.  Now it works, yea!