Panzer: Updating DOFManager to work with blocked assembly
Previously only the FEI version did (not the "standard" DOFManager).
The modification makes it possible to use the standard dof manager
with the block assembly algorithm. I've also modified the unit
tests to excercise this new capability.

Further work on the standard DOFManager needs to be done to clean
up its internal data structure and make sure it is easily modifable
by others.