Fixing testingToll() to work for all scalar types (satisfying OCP) (Trilinos #5861)
The old code violated Item 66 "Don't specialize function templates" and did
not satisfy OCP in that it has a satic list of integral types to handle
integral types not having eps().  This commit uses template conditionals and
partial template specialization correctly and keys off of
ScalarTraits::hasMachineParameters so it will now work for any Scalar type
that has a properly defined ScalarTraits specilaizaiton.

This allows the test code to build with Scalar="unsigned long int".

However, a lot of Tpetra tests are failing with these new unsinged scalar
types.