Adding XML control of verbosity level.
Not sure why we didn't have this before.  I couldn't determine how Belos handles its MsgType
in ParameterLists loaded from XML.  So I just created a std::map with
keys equal to the valid strings, and the equivalent MueLu::MsgType's as the values.  The
verbosity is pulled out of the PL (converted from XML), and if it ain't in the map,
I throw an exception.  Not elegant, but it works.

It would be better to put all this in a class.