Teuchos: Fix issue with backwards compatibility
The change to subpackages caused issues with the windows DLL testing. The fix
for the DLL testing removed a macro that is needed by one of the tests which
caused the backwards compatibility testing to complain. Since we likely aren't
going to do backwards compatibility testing on windows the need to ensure that
this macro name still exists is minimal so I am just defining it as blank to
allow the test to continue building.

Build/Test Cases Summary
Enabled Packages: TeuchosCore
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=1229,notpassed=0 (104.97 min)
1) SERIAL_RELEASE => passed: passed=1186,notpassed=0 (49.78 min)