Tpetra: Distributor: Removed barrier between recv & send by default.
Distributor was putting a barrier between receives and sends by
default.  (The ParameterList default was no barrier, but the
constructor default was to enable the barrier.)  I changed this
so there is no barrier by default.  This should improve performance
of Import and Export operations.

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => passed: passed=44,notpassed=0 (1.50 min)
1) SERIAL_RELEASE => passed: passed=34,notpassed=0 (0.55 min)