Epetra,Teuchos,Kokkos,Tpetra: Fixed some compiler warnings.
Tpetra now builds clean with Clang 3.2 "-Weverything -std=c++11"
(the language specification is necessary to avoid warnings on the
language extension "long long", which is part of C++11).  I haven't
systematically gone through Epetra, Teuchos, or Kokkos in this commit,
but I did fix warnings in all of them.

Teuchos memory management classes turned out to need a bit of work to
build clean of warnings.  In particular, there were silent size_type
to int conversions for debug code ("assertIndex") which were not 64-bit
clean.  This could affect debug checks for very large arrays.  Fixing
them only required interface changes to private methods, so the fix
should not affect user code.

Note that different compilers warn on different things.  I've added
internal documentation of a case where it might not be possible to
please both the Intel compiler and Clang.  I favored the existing
fix for the Intel compiler warning in that case.

Clang 3.2 has the nice feature that it warns on HTML errors in Doxygen
documentation, such as HTML tags that aren't closed.  I took advantage
of this to fix some bugs in the documentation of the Teuchos memory
management classes.  This should improve the generated HTML output.