MueLu: utils: the class generating script now updates CMakeLists.txt
If we develop a new class, we run src/Utils/ClassList/gen_all.sh script
in order to do all the extra work. For a new class, it constructs
several new files. But these files are not being caught in other
people's build directories as CMakeLists.txt is not being updated.
Now we add an extra line to CMakeLists.txt to enforce the rebuild.
A quick and dirty solution, though.