Fixed issue of inconsistant Teuchos::MpiComm tags (Trilinos #5740)
With this commit, I have taken a unit test provided by Stuart Slattery (see
Trilinos #5740) and added it as one of the unit tests for Teuchos::MpiComm.
The only significant way that I changed the unit test code was to allow it to
run with any number of processes.  If there is just one process, it passes by
default.  If there are two or more processes in MPI_COMM_WORLD, it creates a
subcommunicator with just two processes and then does the rest, just as
before.

The way that I fixed the unit test was to change Teuchos::MpiComm so that when
it sets the default tag_ from tagCounter_, it does an MPI_Bacast() from the
root process to ensure consistency.  The even increasing count should take
care of duplicate tages with the same MPI_Comm object.

There is a longer discussion in Triinos #5740 that might have us deprecate
(and remove in Trilinos 12.0) the usage of a default tag argument but we will
have to see.