Fixing the test so that it works correctly in optimized mode (Trilinos #5740)
The test was written to call getSize() and getRank() on the masterComm with
just ranks 0 and 1 on all processes.  In a debug build, this worked just fine
(I don't know how).  In an optimized build, it segfaulted.  I was able to
create an MPI_RELEASE build and reproduce the segfault.  The fix was simply to
not call getSize() and getRank() on all processes (which I should have known).

With this simple change, the test works just fine in an MPI_RELEASE build.
This should make all of the test failures on CDash go away.